refactor(driver): centralize reusable models
